Fix markup of C# foreach for issue579.
In essense, the 'in' keyword is treated as a contextual keyword when it
is strictly a keyword only.  However, as other code may be dependent on this,
it was left alone.  Instead, in pattern_check_core a condition was added
to stop the loop on the 'in' keyword while in the mode FOR_CONDITION (which is activated)
directly before parsing the for initialization (this may want to be changed so the modes
to not switch until after for initialiation is complete).  Once stopped, the remaining code,
should detect if a declaration or not a declartion correctly.  Test cases were
added to verify the correction.