Default to a variable instead of a function declaration when in a function block.
This was added to pattern_check_core right before functions would have been set as the type.
This was needed there because of Java anonymous functions.  Basically, we must be in a function
body, it must be possible a variable, and it ends with a ';'.  I also put it there so, anonymous
functions in other languages could be done correctly if necessary, however, no test cases, were
broken if put before the function checks.